@@ -5,12 +5,20 @@
   "main": "index.js",
   "scripts": {
     "lint": "./node_modules/.bin/eslint ./",
-    "dev": "node ssr-server.js"
+    "build": "next build",
+    "dev": "ts-node-dev --project tsconfig.server.json --respawn --transpileOnly ./src/ssr-server.js",
+    "prod": "tsc --project tsconfig.server.json --rootDir ./src --outDir ./build && node ./build/ssr-server.js",
+    "tsc": "tsc --project tsconfig.server.json --outDir ./build --rootDir ./src"
   },
   "keywords": [],
   "author": "hack4impact",
   "license": "MIT",
   "devDependencies": {
+    "@babel/core": "^7.6.2",
+    "@babel/node": "^7.6.2",
+    "@babel/preset-env": "^7.6.2",
+    "@types/node": "^12.11.1",
+    "@types/react": "^16.9.9",
     "babel-eslint": "^10.0.3",
     "concurrently": "^4.1.2",
     "dotenv": "^8.1.0",
@@ -23,9 +31,11 @@
     "lint-staged": "^9.4.0",
     "mongodb": "^3.3.2",
     "mongoose": "^5.7.3",
+    "nodemon": "^1.19.2",
     "prettier": "1.18.2",
     "sendgrid": "^5.2.3",
-    "typescript": "^3.6.3"
+    "ts-node-dev": "^1.0.0-pre.43",
+    "typescript": "^3.6.4"
   },
   "husky": {
     "hooks": {
@@ -44,8 +54,12 @@
     "@material-ui/styles": "^4.5.0",
     "@sendgrid/mail": "^6.4.0",
     "@zeit/next-css": "^1.0.1",
+    "apollo-server": "^2.9.5",
+    "apollo-server-express": "^2.9.5",
     "express": "^4.17.1",
-    "next": "^9.0.7",
+    "graphql": "^14.5.8",
+    "merge-graphql-schemas": "^1.7.0",
+    "next": "^9.1.1",
     "react": "^16.10.1",
     "react-dom": "^16.10.1"
   }
