@@ -1,5 +1,8 @@
-const express = require("express");
-const next = require("next");
+import graphql from "./backend/graphql";
+import { ApolloServer } from "apollo-server-express";
+
+import express from "express";
+import next from "next";
 
 const dev = process.env.NODE_ENV !== "production";
 const app = next({ dev });
@@ -8,19 +11,30 @@ const handle = app.getRequestHandler();
 // Connect to mongodb using mongoose
 require("dotenv").config();
 const mongoose = require("mongoose");
-mongoose.connect(process.env.MONGODB_URI);
-const models = require("./backend/models/models");
+mongoose.connect(process.env.MONGODB_URI, {
+  useUnifiedTopology: true,
+  useNewUrlParser: true
+});
 
+const models = require("./backend/models");
 const db = mongoose.connection;
+
 db.on("error", console.error.bind(console, "Connection error:"));
 db.once("open", async function() {
   console.log("Connected to database");
+});
+
+const apolloServer = new ApolloServer({
+  typeDefs: graphql.schema,
+  resolvers: graphql.resolvers,
+  playground: true
 });
 
 app
   .prepare()
   .then(() => {
     const server = express();
+    apolloServer.applyMiddleware({ app: server, path: "/api" });
 
     server.get("*", (req, res) => {
       return handle(req, res);
