@@ -2,7 +2,7 @@
 const mongoose = require("mongoose");
 
 // Import schemas
-const UserSchema = require("./user");
+const UserSchema = require("./modules/user/model");
 
 // Create a database model for each schema
 const User = mongoose.model("User", UserSchema);
